Add method to enumerate zfs props.
Add a method to enumerate zfs props from zfs_props.

A method to enumerate zpool props will be added later if it's needed.

In addition, split the lib file into discrete files for
better maintenence.
