+++ libzfs/Cargo.toml
@@ -12,3 +12,4 @@ serde = "1.0"
 serde_derive = "1.0"
 foreign-types = "0.3"
 cstr-argument = "0.0"
+lazy_static = "1.0"
